Add initial but working mail switcher
This is initial, but working version of a script called by
postfix/master in 'catchall' transport. It reads an e-mail entirely to
memory and replace sender based on fake senders table. Also replace
recipient based on aliases table.

In the version an e-mail keeps looping as the recipient is not replaced
by aliases mechanism. SQL function will return proper recipient.